Backend patient service terminer, test unitaire terminer, mise en place des tests d'intégrations

+++ patient-service/src/main/java/com/glucovision/patientservice/service/PatientService.java
@@ -5,6 +5,8 @@ import com.glucovision.patientservice.repository.PatientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -13,19 +15,36 @@ public class PatientService {
     @Autowired
     private PatientRepository patientRepository;
 
+    //retourne une liste de patient
+    public List<Patient> findAll() {
+        return patientRepository.findAll();
+    }
+
     // Ajouter un patient
     public Patient addPatient(Patient patient) {
+        if (patient == null
+                || patient.getFirstName() == null || patient.getFirstName().trim().isEmpty()
+                || patient.getLastName() == null || patient.getLastName().trim().isEmpty()
+                || patient.getBirthDate() == null
+                || patient.getGender() == null) {
+
+            throw new IllegalArgumentException("Invalid patient data");
+        }
+
         return patientRepository.save(patient);
     }
 
     // Trouver un patient par son ID
-    public Optional<Patient> findPatientById(Long id) {
-        return patientRepository.findById(id);
+    public Patient findPatientById(Long id) {
+        return patientRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("Patient with ID " + id + " not found"));
     }
+
 
     // Trouver un patient par son nom
     public Patient findPatientByName(String lastName) {
-        return patientRepository.findByLastName(lastName);
+        return patientRepository.findByLastName(lastName).orElseThrow(()->
+                 new NoSuchElementException ("Patient with name " + lastName + " not found"));
     }
 
     // Mettre à jour un patient
@@ -39,6 +58,9 @@ public class PatientService {
 
     // Supprimer un patient
     public void deletePatient(Long id) {
+        if (!patientRepository.existsById(id)) {
+            throw new NoSuchElementException("Patient with ID " + id + " not found");
+        }
         patientRepository.deleteById(id);
     }
 }
